Restore colorized logging

+++ util/util_bsd.go
@@ -8,9 +8,9 @@ import (
 	"unsafe"
 )
 
-func isTTY(fd int) bool {
+func isTTY(fd uintptr) bool {
 	var termios syscall.Termios
-	_, _, err := syscall.Syscall6(syscall.SYS_IOCTL, uintptr(fd), syscall.TIOCGETA, uintptr(unsafe.Pointer(&termios)), 0, 0, 0)
+	_, _, err := syscall.Syscall6(syscall.SYS_IOCTL, fd, syscall.TIOCGETA, uintptr(unsafe.Pointer(&termios)), 0, 0, 0)
 	return err == 0
 }
